move components out of pages directory and into /components
