+++ packages/prop-house-webapp/src/components/WinningProposalBanner/WinningProposalBanner.tsx
